Moved functions from open_subtitle.go to stl.go
